Decouple models from CMSObject
