+++ libraries/src/Object/CMSObject.php
@@ -24,14 +24,7 @@ namespace Joomla\CMS\Object;
  */
 class CMSObject
 {
-    /**
-     * An array of error messages or Exception objects.
-     *
-     * @var    array
-     * @since  1.7.0
-     * @deprecated  3.1.4  JError has been deprecated
-     */
-    protected $_errors = array();
+    use LegacyErrorHandlingTrait;
 
     /**
      * Class constructor, overridden in descendant classes.
@@ -126,51 +119,6 @@ class CMSObject
     }
 
     /**
-     * Get the most recent error message.
-     *
-     * @param   integer  $i         Option error index.
-     * @param   boolean  $toString  Indicates if Exception objects should return their error message.
-     *
-     * @return  string   Error message
-     *
-     * @since   1.7.0
-     * @deprecated 3.1.4  JError has been deprecated
-     */
-    public function getError($i = null, $toString = true)
-    {
-        // Find the error
-        if ($i === null) {
-            // Default, return the last message
-            $error = end($this->_errors);
-        } elseif (!\array_key_exists($i, $this->_errors)) {
-            // If $i has been specified but does not exist, return false
-            return false;
-        } else {
-            $error = $this->_errors[$i];
-        }
-
-        // Check if only the string is requested
-        if ($error instanceof \Exception && $toString) {
-            return $error->getMessage();
-        }
-
-        return $error;
-    }
-
-    /**
-     * Return all errors, if any.
-     *
-     * @return  array  Array of error messages.
-     *
-     * @since   1.7.0
-     * @deprecated 3.1.4  JError has been deprecated
-     */
-    public function getErrors()
-    {
-        return $this->_errors;
-    }
-
-    /**
      * Modifies a property of the object, creating it if it does not already exist.
      *
      * @param   string  $property  The name of the property.
@@ -211,20 +159,5 @@ class CMSObject
         }
 
         return false;
-    }
-
-    /**
-     * Add an error message.
-     *
-     * @param   string  $error  Error message.
-     *
-     * @return  void
-     *
-     * @since   1.7.0
-     * @deprecated 3.1.4  JError has been deprecated
-     */
-    public function setError($error)
-    {
-        $this->_errors[] = $error;
     }
 }
